requirements.txt
django==4.2.*
django-image-cropping>1
easy-thumbnails>2
environs[django]>=10
